pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sf-cleaner"
version = "0.1.0"
description = "Salesforce lead enrichment via LinkedIn + ZoomInfo using LangGraph, vector cache, and local LLM"
requires-python = ">=3.11"

dependencies = [
    # Salesforce
    "simple-salesforce>=1.12",

    # LangGraph + LangChain
    "langgraph>=0.2",
    "langchain>=0.2",
    "langchain-community>=0.2",
    "langchain-ollama>=0.1",

    # Vector DB
    "pinecone-client>=3.0",
    "pgvector>=0.2",
    "psycopg2-binary>=2.9",

    # Embeddings
    "openai>=1.0",           # text-embedding-3-small via OpenAI SDK

    # Task queue
    "celery>=5.3",
    "redis>=5.0",

    # API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",

    # Data / validation
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "httpx>=0.27",

    # Utilities
    "python-dotenv>=1.0",
    "structlog>=24.1",
    "tenacity>=8.3",
]

[project.optional-dependencies]
airflow = [
    "apache-airflow>=2.9",
    "apache-airflow-providers-celery>=3.6",
]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
